Increase default CPU limits for cert-exporter
We've observed heavy CPU throttling with the default limits on multiple
clusters. For the secrets exporter, the CPU throttling had spikes of up to
0.2 seconds / scheduled second, which led to frequent TargetDown alerts
in Prometheus with the default scrape timeout of 10s.

After updating the CPU limit for the secrets exporter to 500m manually
on one affected cluster, throttling consistently stays below
0.02 seconds / scheduled second, and we haven't had any TargetDown alerts
since adjusting the limit.

This commit updates the default CPU limits for the secrets and hostpaths
exporters to 500m and 250m respectively, increasing both limits by a
factor 2.5. We keep the memory requests and limits, and the CPU
requests, since those are generally reasonable looking at the resource
consumption of the pods.