Fix default CPU limit for cert-exporter v2.13.0
After determining a good CPU limit to avoid excessive throttling for
cert-exporter v2.12.1 (cf. #24), we also merged #8 to update the
cert-exporter container image to v2.13.0. However, it turns out that the
CPU requirements for v2.13.0 are completely different to v2.12.1.

After some more experimentation we've determined that 1000m seems to be
a default CPU limit for v2.13.0 where we avoid excessive throttling even
on clusters with larger amounts of certificates.